Move tzLocal, fzLocal, gradientScenes, and knownEffects to lib
